+++ src/transmission/mod.rs
@@ -1,32 +1,55 @@
 //! Module contains sprakpost email sending api
 //!
 //! ### Example
-//! ```rust
+//!  ```rust
+//!  extern crate sparkpost;
 //!
+//! use sparkpost::transmission::{Transmission, Message, EmailAddress, TransmissionResponse};
 //!
-//! use sparkpost::transmission::{Message, EmailAddress, Transmission, TransmissionResponse};
-//! let tm = Transmission::new("api_key_form_env".to_string(),
-//!                           "https://api.eu.sparkpost.com/api/v1/transmissions".into());
+//! let tm = Transmission::new("api_key".to_string(), "https://api.eu.sparkpost.com/api/v1".to_string());
+//! let mut email: Message = Message::new(
+//!                              EmailAddress::new("marketing@example.sink.sparkpostmail.com", "Example Company")
+//!                          );
 //!
-//! let mut email = Message::new(EmailAddress::with_name("marketing@example.sink.sparkpostmail.com", "Example Company"));
-//! email.add_recipient("wilma@example.sink.sparkpostmail.com".into())
-//!        .campaign_id("postman_inline_both_example")
-//!        .subject("SparkPost inline template example")
-//!        .html("<html><body>Here is your inline html, {{first_name or 'you great person'}}!<br></body></html>")
-//!        .text("Here is your plain text, {{first_name or 'you great person'}}!");
-//! // send message
-//! // returns result with Transmission Response or reqwest::error::Error
-//! let response: Result<TransmissionResponse, _> = tm.send(&email);
+//! email.add_recipient("name@domain.com".into())
+//!      .subject("My Awesome email 😎")
+//!      .html("<h1>html body of the email</h1>")
+//!      .text("text body of the email");
+//!
+//! let result = tm.send(&email);
+//!
+//! match result {
+//!    Ok(res) => {
+//!         println!("{:?}", &res);
+//!         match res {
+//!             TransmissionResponse::ApiResponse(api_res) => {
+//!              //   assert_eq!(1, api_res.total_accepted_recipients);
+//!              //   assert_eq!(0, api_res.total_rejected_recipients);
+//!             }
+//!             TransmissionResponse::ApiError(errors) => {
+//!                 println!("res: \n {:#?}", &errors);
+//!             }
+//!         }
+//!     }
+//!     Err(error) => {
+//!         println!("error \n {:#?}", error);
+//!     }
+//! }
 //!
 //! ```
 use reqwest::{
     header::{HeaderMap, HeaderValue, ACCEPT, AUTHORIZATION, CONTENT_TYPE},
-    Client, Error as ReqError,
+    Client, Error,
 };
 
 mod message;
+mod recipients;
 
 pub use self::message::*;
+pub use self::recipients::*;
+
+/// Reqwest Error
+pub type ReqError = Error;
 
 /// Transmission result returned by the API
 ///
@@ -114,10 +137,8 @@ mod tests {
             get_api_key(),
             "https://api.eu.sparkpost.com/api/v1/transmissions".into(),
         );
-        let mut email: Message = Message::new(EmailAddress::with_name(
-            "hello@email.letsorganise.app",
-            "noreply",
-        ));
+        let mut email: Message =
+            Message::new(EmailAddress::new("hello@email.letsorganise.app", "noreply"));
         email
             .add_recipient("test@hgill.io".into())
             .subject("Testing builder email sandbox")